+++ crates/full-node/sov-sequencer/tests/integration/main.rs
@@ -1,5 +1,4 @@
 mod preferred_end_to_end;
-mod preferred_without_node;
 mod rest_api;
 mod std_batch_builder;
 mod thin_sequencer;